+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.12.9] - in progress
 
 - New function ``streamer.tee``.
+- New class ``streamer.IterProcessQueue``.
 
 
 ## [0.12.8] - 2023-05-17